requirements.txt
asgiref==3.11.0
babel==2.17.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==4.2.27
django-countries==8.2.0
django-environ==0.12.0
django-extra-views==0.16.0
django-haystack==3.3.0
django-localflavor==3.1
django-oscar==4.1
django-phonenumber-field==8.4.0
django-tables2==2.7.0
django-treebeard==4.8.0
django-widget-tweaks==1.5.0
easy-thumbnails==2.10.1
factory_boy==3.3.3
Faker==38.2.0
idna==3.11
packaging==25.0
paypalrestsdk==1.13.3
phonenumbers==9.0.20
pillow==12.0.0
purl==1.6
pycountry==24.6.1
pycparser==2.23
pyOpenSSL==25.3.0
python-stdnum==2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
whitenoise==6.11.0
Whoosh==2.7.4
